fix: upgrade pino (#311)
