+++ .gitignore
@@ -1,4 +1,4 @@
 .nyc_output
+.tap
 coverage
 node_modules
-.tap
