+++ package.json
@@ -24,8 +24,7 @@
   "license": "MIT",
   "repository": "github:probot/pino",
   "devDependencies": {
-    "@types/pino": "^7.0.5",
-    "pino": "^6.6.0",
+    "pino": "^9.3.2",
     "prettier": "^3.4.2",
     "tap": "^21.0.1"
   },
